refactor: Member와 관련된 Filter 기능을 Filter API로 이전
