+++ backend/src/main/java/com/carffeine/carffeine/car/domain/MemberCarRepository.java
@@ -1,5 +1,6 @@
-package com.carffeine.carffeine.member.domain;
+package com.carffeine.carffeine.car.domain;
 
+import com.carffeine.carffeine.member.domain.Member;
 import org.springframework.data.repository.Repository;
 
 import java.util.Optional;
