+++ backend/src/main/java/com/carffeine/carffeine/car/service/CarService.java
@@ -4,13 +4,13 @@ import com.carffeine.carffeine.admin.exception.AdminException;
 import com.carffeine.carffeine.admin.exception.AdminExceptionType;
 import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.CarRepository;
+import com.carffeine.carffeine.car.domain.MemberCar;
+import com.carffeine.carffeine.car.domain.MemberCarRepository;
 import com.carffeine.carffeine.car.exception.CarException;
 import com.carffeine.carffeine.car.exception.CarExceptionType;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.car.service.dto.CarsRequest;
 import com.carffeine.carffeine.member.domain.Member;
-import com.carffeine.carffeine.member.domain.MemberCar;
-import com.carffeine.carffeine.member.domain.MemberCarRepository;
 import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.member.exception.MemberException;
 import com.carffeine.carffeine.member.exception.MemberExceptionType;
