+++ backend/src/test/java/com/carffeine/carffeine/car/service/CarServiceTest.java
@@ -4,16 +4,16 @@ import com.carffeine.carffeine.admin.exception.AdminException;
 import com.carffeine.carffeine.admin.exception.AdminExceptionType;
 import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.CarRepository;
+import com.carffeine.carffeine.car.domain.MemberCar;
+import com.carffeine.carffeine.car.domain.MemberCarRepository;
 import com.carffeine.carffeine.car.exception.CarException;
 import com.carffeine.carffeine.car.exception.CarExceptionType;
 import com.carffeine.carffeine.car.fake.FakeCarRepository;
+import com.carffeine.carffeine.car.fake.FakeMemberCarRepository;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.car.service.dto.CarsRequest;
-import com.carffeine.carffeine.member.domain.FakeMemberCarRepository;
 import com.carffeine.carffeine.member.domain.FakeMemberRepository;
 import com.carffeine.carffeine.member.domain.Member;
-import com.carffeine.carffeine.member.domain.MemberCar;
-import com.carffeine.carffeine.member.domain.MemberCarRepository;
 import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.member.domain.MemberRole;
 import org.assertj.core.api.AssertionsForClassTypes;
