+++ backend/src/test/java/com/carffeine/carffeine/filter/controller/FilterControllerTest.java
@@ -3,9 +3,12 @@ package com.carffeine.carffeine.filter.controller;
 import com.carffeine.carffeine.filter.controller.dto.FiltersResponse;
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.filter.domain.FilterType;
+import com.carffeine.carffeine.filter.domain.MemberFilter;
 import com.carffeine.carffeine.filter.service.dto.FilterRequest;
 import com.carffeine.carffeine.filter.service.dto.FiltersRequest;
 import com.carffeine.carffeine.helper.MockBeanInjection;
+import com.carffeine.carffeine.member.domain.Member;
+import com.carffeine.carffeine.member.fixture.MemberFixture;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -209,5 +212,82 @@ class FilterControllerTest extends MockBeanInjection {
                                 fieldWithPath("connectorTypes[0]").type(JsonFieldType.ARRAY).description("충전기 타입")
                         )
                 ));
+    }
+
+    @Test
+    void 회원의_모든_필터를_조회한다() throws Exception {
+        // given
+        List<Filter> filters = List.of(
+                Filter.of("충전소 회사", FilterType.COMPANY.getName()),
+                Filter.of("충전소 회사2", FilterType.COMPANY.getName()),
+                Filter.of("2", FilterType.CAPACITY.getName()),
+                Filter.of("DC_COMBO", FilterType.CONNECTOR_TYPE.getName())
+        );
+
+        // when
+        when(filterService.findMemberFilters(any(), any())).thenReturn(filters);
+
+        // then
+        mockMvc.perform(RestDocumentationRequestBuilders.get("/members/{memberId}/filters", 1L)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer token~~")
+                ).andExpect(status().isOk())
+                .andDo(customDocument("find_member_filters",
+                                requestHeaders(headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")),
+                                pathParameters(parameterWithName("memberId").description("Member id")),
+                                responseFields(
+                                        fieldWithPath("companies[0]").type(JsonFieldType.ARRAY).description("충전기 회사"),
+                                        fieldWithPath("capacities[0]").type(JsonFieldType.ARRAY).description("충전 용량"),
+                                        fieldWithPath("connectorTypes[0]").type(JsonFieldType.ARRAY).description("충전기 타입")
+                                )
+                        )
+                );
+    }
+
+    @Test
+    void 회원의_선호_필터를_등록한다() throws Exception {
+        // given
+        Member member = MemberFixture.일반_회원;
+
+        FiltersRequest filtersRequest = new FiltersRequest(
+                List.of(
+                        new FilterRequest(FilterType.COMPANY.getName(), "충전소 회사"),
+                        new FilterRequest(FilterType.CAPACITY.getName(), "2"),
+                        new FilterRequest(FilterType.CONNECTOR_TYPE.getName(), "DC_COMBO")
+                )
+        );
+
+        List<MemberFilter> memberFilters = List.of(
+                new MemberFilter(member, Filter.of("충전소 회사", FilterType.COMPANY.getName())),
+                new MemberFilter(member, Filter.of("2", FilterType.CAPACITY.getName())),
+                new MemberFilter(member, Filter.of("DC_COMBO", FilterType.CONNECTOR_TYPE.getName()))
+        );
+
+        // when
+        when(filterService.addMemberFilters(any(), any(), any())).thenReturn(memberFilters);
+
+        // then
+        mockMvc.perform(RestDocumentationRequestBuilders.post("/members/{memberId}/filters", 1L)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer token~~")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(filtersRequest))
+                ).andExpect(status().isOk())
+                .andDo(customDocument("add_member_filters",
+                                requestHeaders(headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")),
+                                pathParameters(parameterWithName("memberId").description("Member id")),
+                                requestFields(
+                                        fieldWithPath("filters[0].type").type(JsonFieldType.STRING).description("필터 종류"),
+                                        fieldWithPath("filters[0].name").type(JsonFieldType.STRING).description("필터 이름"),
+                                        fieldWithPath("filters[1].type").type(JsonFieldType.STRING).description("필터 종류"),
+                                        fieldWithPath("filters[1].name").type(JsonFieldType.STRING).description("필터 이름"),
+                                        fieldWithPath("filters[2].type").type(JsonFieldType.STRING).description("필터 종류"),
+                                        fieldWithPath("filters[2].name").type(JsonFieldType.STRING).description("필터 이름")
+                                ),
+                                responseFields(
+                                        fieldWithPath("companies[0]").type(JsonFieldType.ARRAY).description("충전기 회사"),
+                                        fieldWithPath("capacities[0]").type(JsonFieldType.ARRAY).description("충전 용량"),
+                                        fieldWithPath("connectorTypes[0]").type(JsonFieldType.ARRAY).description("충전기 타입")
+                                )
+                        )
+                );
     }
 }
