+++ backend/src/test/java/com/carffeine/carffeine/filter/service/FilterServiceTest.java
@@ -10,13 +10,19 @@ import com.carffeine.carffeine.car.fake.FakeCarRepository;
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.filter.domain.FilterRepository;
 import com.carffeine.carffeine.filter.domain.FilterType;
+import com.carffeine.carffeine.filter.domain.MemberFilter;
+import com.carffeine.carffeine.filter.domain.MemberFilterRepository;
 import com.carffeine.carffeine.filter.fake.FakeFilterRepository;
+import com.carffeine.carffeine.filter.fake.FakeMemberFilterRepository;
 import com.carffeine.carffeine.filter.service.dto.FilterRequest;
 import com.carffeine.carffeine.filter.service.dto.FiltersRequest;
 import com.carffeine.carffeine.member.domain.FakeMemberRepository;
 import com.carffeine.carffeine.member.domain.Member;
 import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.member.domain.MemberRole;
+import com.carffeine.carffeine.member.exception.MemberException;
+import com.carffeine.carffeine.member.exception.MemberExceptionType;
+import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -34,8 +40,9 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 class FilterServiceTest {
 
     private FilterService filterService;
-    private FilterRepository filterRepository;
     private MemberRepository memberRepository;
+    private FilterRepository filterRepository;
+    private MemberFilterRepository memberFilterRepository;
     private CarRepository carRepository;
     private CarFilterRepository carFilterRepository;
 
@@ -44,11 +51,12 @@ class FilterServiceTest {
 
     @BeforeEach
     void setup() {
-        filterRepository = new FakeFilterRepository();
         memberRepository = new FakeMemberRepository();
+        filterRepository = new FakeFilterRepository();
+        memberFilterRepository = new FakeMemberFilterRepository();
         carRepository = new FakeCarRepository();
         carFilterRepository = new FakeCarFilterRepository();
-        filterService = new FilterService(filterRepository, memberRepository, carRepository, carFilterRepository);
+        filterService = new FilterService(memberRepository, filterRepository, memberFilterRepository, carRepository, carFilterRepository);
 
         admin = memberRepository.save(Member.builder()
                 .memberRole(MemberRole.ADMIN)
@@ -147,5 +155,60 @@ class FilterServiceTest {
         assertThatThrownBy(() -> filterService.addCarFilters(member.getId(), savedCar.getId(), request))
                 .isInstanceOf(AdminException.class)
                 .hasMessage(AdminExceptionType.NOT_ADMIN.message());
+    }
+
+    @Test
+    void 회원이_등록한_모든_필터를_찾는다() {
+        // given
+        List<MemberFilter> memberFilters = memberFilterRepository.saveAll(
+                List.of(new MemberFilter(member, Filter.of("DC_COMBO", FilterType.CONNECTOR_TYPE.getName())))
+        );
+
+        // when
+        List<Filter> result = filterService.findMemberFilters(member.getId(), member.getId());
+
+        // then
+        assertThat(memberFilters.get(0).getFilter()).isEqualTo(result.get(0));
+    }
+
+    @Test
+    void 다른_회원의_필터를_조회하면_예외를_반환한다() {
+        // when & then
+        AssertionsForClassTypes.assertThatThrownBy(() -> filterService.findMemberFilters(member.getId() + 1, member.getId()))
+                .isInstanceOf(MemberException.class)
+                .hasMessage(MemberExceptionType.INVALID_ACCESS.message());
+    }
+
+    @Test
+    void 회원이_선호하는_필터를_등록한다() {
+        // given
+        filterRepository.saveAll(List.of(
+                Filter.of("충전소 회사", FilterType.COMPANY.getName()),
+                Filter.of("2.00", FilterType.CAPACITY.getName()),
+                Filter.of("DC_COMBO", FilterType.CONNECTOR_TYPE.getName())
+        ));
+
+        FiltersRequest filtersRequest = new FiltersRequest(
+                List.of(
+                        new FilterRequest(FilterType.COMPANY.getName(), "충전소 회사"),
+                        new FilterRequest(FilterType.CAPACITY.getName(), "2.00"),
+                        new FilterRequest(FilterType.CONNECTOR_TYPE.getName(), "DC_COMBO")
+                )
+        );
+
+        // when
+        List<MemberFilter> result = filterService.addMemberFilters(member.getId(), member.getId(), filtersRequest);
+
+        // then
+        List<MemberFilter> memberFilters = memberFilterRepository.findAllByMember(member);
+        assertThat(result).hasSize(memberFilters.size());
+    }
+
+    @Test
+    void 다른_회원의_필터를_등록하면_예외를_반환한다() {
+        // when & then
+        AssertionsForClassTypes.assertThatThrownBy(() -> filterService.addMemberFilters(member.getId() + 1, member.getId(), null))
+                .isInstanceOf(MemberException.class)
+                .hasMessage(MemberExceptionType.INVALID_ACCESS.message());
     }
 }
